Broke up two large line functions into 2 files.
Broke up the larger functions into smaller pieces for the norm. Still
not fully up to norm. Need to add comments for functions and do testing.
