+++ includes/engine.h
@@ -6,7 +6,7 @@
 /*   By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/23 10:24:47 by dpentlan          #+#    #+#             */
-/*   Updated: 2023/11/29 15:55:06 by dpentlan         ###   ########.fr       */
+/*   Updated: 2023/11/30 08:47:19 by dpentlan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -154,6 +154,11 @@ bool	ft_draw_player(t_engine *engine, t_map_data *map_data,
 void	ft_move_player(t_player *player, t_direction direction);
 void	ft_rotate_player(t_player *player, t_direction direction);
 bool	ft_draw_rays(t_engine *engine, t_player *player, t_map_data *map_data);
+void	ft_horiz_check(t_player *player, t_map_data *map_data, t_ray *h_ray,
+				int max_dof);
+void	ft_vert_check(t_player *player, t_map_data *map_data, t_ray *v_ray,
+				int max_dof);
+double	ft_distance(double pt1x, double pt1y, double pt2x, double pt2y);
 		//***** images *****/
 bool	ft_image_init(t_image *image,
 			void *mlx_ptr, t_image_id ID, t_point size);
